Fist Commit.
FFMPEG is a hard requirement for these programs. It also needs to be in your PATH variable so that it can be called by only typing 'ffmpeg'

Also uses YoutubeExplode, Cli.Wrap, and Newtonsoft.JSON

Adds a downloader project which can download audio and video and can mux with the help of FFMPEG.

Adds a playlist downloader project that uses the downloader exe, it must also be in your path variable.